HttpResponse separate file & refactoring
